change type of number to int + test

+++ src/main/java/org/esgi/el_presidente/core/ressources/Agriculture.java
@@ -1,27 +1,27 @@
 package org.esgi.el_presidente.core.ressources;
 
 public class Agriculture implements IslandPart {
-  private double partOfIsland;
-  private final double foodProductionCoefficient;
+  private int partOfIsland;
+  private final int foodProductionCoefficient;
 
-  public Agriculture(double partOfIsland, double foodProductionCoefficient) {
+  public Agriculture(int partOfIsland, int foodProductionCoefficient) {
     this.partOfIsland = partOfIsland;
     this.foodProductionCoefficient = foodProductionCoefficient;
   }
 
-  public double yearlyProductionOfFood() {
+  public int yearlyProductionOfFood() {
     return partOfIsland * foodProductionCoefficient;
   }
 
-  public void expand(double additionalSize) {
+  public void expand(int additionalSize) {
     partOfIsland += additionalSize;
   }
 
-  public double getSize() {
+  public int getSize() {
     return partOfIsland;
   }
 
-  public void shrink(double deductSize) {
+  public void shrink(int deductSize) {
     partOfIsland -= deductSize;
   }
 }
